Updated files for the Team 3 dealership project
These files reflect the changes that I mentioned in the following post I uploaded to the Team 3 IST242 Microsoft Teams chat:

Pardon me for the late message. I've been working on the code for our group project all night.. What I have working so far is:

A set of test cars for the car arraylist
Most of two sample dealership objects for the dealership arrayList (need to fix manager and formatting)
A login/create an account prompt that leads to a functional account creation process.
A login process that verifies User name entries and password entries
A user menu with the customer functions mostly done (I need to fix the financing options prompt as well as add additional functionality for sales associates and management after checking their role typeID)

 

What still needs to be done:
Functions for sales associate processes
Functions for prompting and allowing a customer to add their email address etc. and link it to their user account.

-Quenten Calvano